settings page refactor. everything is working except the address at this point
